Move entries-related files to container level
